created app and model
