+++ backend/generator/profile/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class AccountConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'account'
+    name = 'profile'
